+++ plone/app/event/ical/exporter.py
@@ -423,6 +423,6 @@ class EventsICal(BrowserView):
             'Content-Disposition',
             'attachment; filename="{0}"'.format(name)
         )
-        self.request.response.setHeader('Content-Length', len(ical))
         self.request.response.setHeader('Pragma', 'no-cache')
+        self.request.response.setHeader('Content-Length', len(ical))
         self.request.response.write(ical)